requirements.txt
aiohttp>=3.9.*
click~=8.1.7
python-dotenv~=1.0.1
elasticsearch~=8.12.1
elasticsearch-dsl~=8.12.0
Telethon~=1.34.0
pycryptodome>=3.20.0
aiosignal>=1.3.0
aiofiles>=23.0.0
pandas~=2.2.1
pycryptodome~=3.20.0
google-generativeai>=0.5.0
dacite~=1.8.1
pytz~=2024.1
tqdm~=4.66.2
openai~=1.20.0
colorama~=0.4.6
numpy~=1.26.4